Move casting of floats into loader, fail in TrackGraph on non-int segmentation

+++ tests/matchers/test_ctc.py
@@ -17,8 +17,12 @@ class TestCTCMatcher:
         # shapes don't match
         with pytest.raises(ValueError):
             self.matcher.compute_mapping(
-                TrackingGraph(nx.DiGraph(), segmentation=np.zeros((5, 10, 10))),
-                TrackingGraph(nx.DiGraph(), segmentation=np.zeros((5, 10, 5))),
+                TrackingGraph(
+                    nx.DiGraph(), segmentation=np.zeros((5, 10, 10), dtype=np.uint16)
+                ),
+                TrackingGraph(
+                    nx.DiGraph(), segmentation=np.zeros((5, 10, 5), dtype=np.uint16)
+                ),
             )
 
     def test_end_to_end(self):
